% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_biological_pathways_from_genes.R
\name{enrich.biopathways}
\alias{enrich.biopathways}
\title{Biological pathway enrichment}
\usage{
enrich.biopathways(comparison)
}
\arguments{
\item{comparison}{the comparison name used in the result folder}
}
\description{
Find biological pathway enrichment from genes localized in modified regions
}
