% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_genes_in_regions.R
\name{gene.annotation}
\alias{gene.annotation}
\title{Gene detection}
\usage{
gene.annotation(ensembldb.genetype, comparison)
}
\arguments{
\item{ensembldb.genetype}{the database containing gene name and gene positions}

\item{comparison}{the comparison name used in the result folder}
}
\description{
Find genes in regions of interest
}
