% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{load.data}
\alias{load.data}
\title{Load data}
\usage{
load.data(name.experiment)
}
\arguments{
\item{name.experiment}{the name of a condition}
}
\description{
recover values of replication timing profile for each condition of interest
}
