% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_nonmodified_regions.R
\name{non.modified.regions}
\alias{non.modified.regions}
\title{Non-modified regions generation}
\usage{
non.modified.regions(differential.result, size.chr.table, comparison)
}
\arguments{
\item{differential.result}{the dataframe containing modified regions between conditions of interest}

\item{size.chr.table}{the dataframe containing the size of chromosomes}

\item{comparison}{the comparison name used in the result folder}
}
\description{
Method to recover non-modified regions of the same size as modified regions
}
