% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{replicate.merging}
\alias{replicate.merging}
\title{Merge samples}
\usage{
replicate.merging(list.replicates)
}
\arguments{
\item{list.replicates}{a list containing dataframes of replicate samples}
}
\description{
Merge replicates for a condition of interest
}
