% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_targets-tf_modified_regions.R
\name{targets.tf.enrichment}
\alias{targets.tf.enrichment}
\title{Transcription factor enrichment analysis}
\usage{
targets.tf.enrichment(comparison, tf, organism)
}
\arguments{
\item{comparison}{the name of the differential analysis performed between profiles}

\item{tf}{all transcription factors known for an organism}

\item{organism}{the name of the organism}
}
\description{
Perform a fisher test to find the transcription factors with the most gene targets in modified regions
}
