% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{modified.regions.detection}
\alias{modified.regions.detection}
\title{Differential analysis of replication timing profiles}
\usage{
modified.regions.detection(
  list.cond,
  cond1,
  cond2,
  cond1.loess,
  cond2.loess,
  pval,
  per.dist.detect,
  per.dist.elong,
  comparison,
  chromosomes
)
}
\arguments{
\item{list.cond}{a list of all conditions of the study}

\item{cond1}{the first condition (values from quantile normalization)}

\item{cond2}{the second condition (values from quantile normalization)}

\item{cond1.loess}{the first condition (values from loess smoothing)}

\item{cond2.loess}{the second condition (values from loess smoothing)}

\item{pval}{the p-value threshold}

\item{per.dist.detect}{the distance at which potential modified regions are detected (percentage between 0 and 100)}

\item{per.dist.elong}{the distance at which the elongation of the regions stops (percentage between 0 and 100)}

\item{comparison}{the name of comparison to build result folder}

\item{chromosomes}{the chromosomes studied for the organism of interest}
}
\description{
Determine modified regions between conditions of interest
}
